# Re-encode chunks that are still being appended to when snapshoting.