fixes: https://github.com/prometheus/tsdb/issues/518
Chunks that are still being appended to might return corrupted data when
calling `chunk.Bytes()` that is why incomplete chunks should never be
persisted. This is relevant mainly when snapshotting with the head
included.

The current snapshot test is sufficient to ensure that incomplete chunks(being appended to) are ignored.

Had this idea while reviewing https://github.com/prometheus/tsdb/pull/639