@gouthamve 2019-06-27
But this is an incomplete snapshot then? We can basically not persist 2hrs of ingested data with this.