@brian-brazil 2019-06-27
It's not safe to do this, you need a clean break in the data. If a user wanted this they can skip the head block.