@krasi-georgiev 2019-06-27
Why not safe? this is skipping only few of the last samples not the entire head.

I looked at the code quite a bit and otherwise `chunk.Bytes()` could always return corrupt data.