@brian-brazil 2019-06-27
> Why not safe? this is skipping only few of the last samples not the entire head.

It's randomly skipping different amounts of data for different series.