@krasi-georgiev 2019-06-27
I guess we can run some additional check to ensure all chunks are cut at the same time.
So i understand the problem with different timings Can you give an example of a query that will cause issues?

Do you have any other idea to solve that race?