@bwplotka 2019-06-27
Yes. Reencode those moving chunks: https://github.com/prometheus/tsdb/pull/639