@krasi-georgiev 2019-06-27
Did you see my comment in your pr? I think it still has a race there.