@krasi-georgiev 2019-06-27
Maybe we can combine both prs. Which is
To Reincode all incomplete chunks(head chunks)