@bwplotka 2019-06-27
I am not saying my PR is perfect. Even CI is saying it's not but my main
focus is on remote read TBH so joining forces would be nice. I can push my
changes to some tsdb Branch for easier collaboration @Krasi (:

On Thu, Jun 27, 2019, 15:50 Krasi Georgiev <notifications@github.com> wrote:

> Maybe we can combine both prs. Which is
> To Reincode all incomplete chunks(head chunks)
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/tsdb/pull/641?email_source=notifications&email_token=ABVA3OZCWMBRJNAWTYTKKO3P4THVFA5CNFSM4H32UOK2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODYXLXYY#issuecomment-506379235>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABVA3O3VPX5Q3FW2SWCBTV3P4THVFANCNFSM4H32UOKQ>
> .
>