@codesome 2019-06-28
I had a look at this and #639 , and for some reason I find re-encoding the last chunk safely and using it seems to be a better option than ignoring the last chunk. Reasons would be (1) What @brian-brazil [said](https://github.com/prometheus/tsdb/pull/641#issuecomment-506321857) (2) Snapshot as much data as possible.

#639 still needs to address the race as described by @krasi-georgiev  [here](https://github.com/prometheus/tsdb/pull/639#issuecomment-506300327).