@bwplotka 2019-06-28
Another idea would be to pass somehow else that it is a head block and treat those differently in terms of maxTime being outside of desired meta.MaxTime