@bwplotka 2019-06-28
No worries. Essentially we are now marking `open chunk` in certain way and we rewrite only them. This part makes sense. But to fix edge case we need something more. I propose to let's say:
 
1. leave this marker we have now for `open` blocks. Rewrite those 100%.
2. Add some another way to tell if the chunk is part of head block. Another meta field just for purpose of populate is rather no go (will be hard as it is part of populateBlock and IndexReader interface..), but maybe literally some flag to `compute.write` ? 

Now we will have:
* check if head, open block. If yes rewrite.
* else check if it is outside of meta.Max.
  * if the chunk is part of head and is closed: rewrite
  * anything else: fail.