@krasi-georgiev 2019-06-28
hm still a bit unclear, but IIUC you mean check if a chunk belongs to the head and rewrite when it is open or outside `meta.Max` ? for all other chunks just fail when outside `meta.Maxt`