@bwplotka 2019-07-03
Today is release day and this PR is ready with only one nit: `safeChunk` casting to tell if it's chunk block or not. 

I would add `TODO` and otherwise merge it and do TSDB update on Prometheus side. Any attempt of refactoring `populate` will probably take *days* to do and review, so I would fix this nit later.

Thoughts? @brancz @brian-brazil @codesome @krasi-georgiev