@codesome 2019-07-03
Increase in allocs (and also B/op) are alarming for a small change in `populateBlock`