@codesome 2019-07-03
And confirmed that it is in `populateBlock`, I am investigating it.